xperiment/requirements/base.txt
boto3==1.4.8
Django==1.11.7
django-allauth==0.25.2
django-bootstrap3==8.2.3
django-cors-middleware==1.3.1
django-el-pagination==3.1.0
django-ipware==1.1.6
django-model-utils==3.0.0
django-picklefield==0.3.2
djrill==2.1.0
Fabric==1.13.2
pillow==4.1.1
python-dateutil==2.6.1
python-slugify==1.2.4
pytz==2017.2
requests==2.18.4
rollbar==0.13.11
zappa==0.45.1
unicodecsv==0.14.1
awesome-slugify==1.6.5
django-admin-row-actions==0.0.2
django-ses-boto3==0.1.0
psycopg2==2.7.1
django-storages==1.6.3
brotlipy==0.7.0
django-debug-toolbar==1.9.1
zappa-django-utils==0.3.1